Move ScanInterfaces to cmd/wifi and change the Wifi interface
