+++ pkg/wifi/stub.go
@@ -1,13 +1,8 @@
 package wifi
 
 type StubWifiWorker struct {
-	ScanInterfacesOut  []string
 	ScanWifiOut        []WifiOption
 	ScanCurrentWifiOut string
-}
-
-func (stub StubWifiWorker) ScanInterfaces() ([]string, error) {
-	return stub.ScanInterfacesOut, nil
 }
 
 func (stub StubWifiWorker) ScanWifi() ([]WifiOption, error) {
